Start on link quotas

+++ src/db/modules/link/remove.js
@@ -1,6 +1,7 @@
 const Link = require("../../models/link");
+const Account = require("../../models/account");
 
-module.exports = async ({ ids, account }) => {
+module.exports = async ({ ids, all, account }) => {
     const links = await Link.find({
         id: {
             $in: ids,
@@ -19,12 +20,48 @@ module.exports = async ({ ids, account }) => {
         }
     }
     try {
+        const authorQuotaChanges = await Link.aggregate([
+            {
+                $match: {
+                    id: { $in: ids },
+                },
+            },
+            {
+                $group: {
+                    _id: "$author",
+                    linkCount: { $sum: -1 },
+                    links: { $push: "$id" },
+                },
+            },
+        ]);
+
+        const authorQuotasChanged = authorQuotaChanges.map(async (author) => {
+            const updated = await Account.updateOne({ id: author._id }, { $inc: { "quota.links.used": author.linkCount } }, { new: true });
+            author.success = updated.modifiedCount === 1;
+            return author;
+        });
+
+        const failedAuthorQuotaUpdates = authorQuotasChanged.filter((author) => author.success === false);
+
+        if (failedAuthorQuotaUpdates.length !== 0) {
+            return [null, {
+                code: 500,
+                message: `Internal server error updating link quotas for ${failedAuthorQuotaUpdates.length} unique users, no actions have been performed`,
+            }];
+        }
+
         await Link.deleteMany({
             id: {
                 $in: ids,
             },
         });
-        return [true, null];
+        return [{
+            success: true,
+            message: `Successfully deleted ${ids.length} links`,
+            result: {
+                linksDeleted: ids.length,
+            },
+        }, null];
     } catch (e) {
         console.log(e);
         return [
